remove #tag-stuff include
This include is referencing something which is not defined in the
`<key>repository</key>` section of the file. This causes it to fail validation
when e.g. converting from the tmLanguage file to a .sublime-syntax file.

I think removing it should be harmless given it is referencing something
non-existent anyway.